requirements.txt
pytorch-lightning~=1.5.0
datasets~=2.1.0
transformers~=4.15.0
torchmetrics~=0.7.0
git+https://github.com/ddkang/loss_dropper.git